+++ hyperion/filter/filter.py
