--- hyperion/filters/__init__.py
@@ -1,1 +1,0 @@
-from .filters import Filter
