+++ setup.py
@@ -96,8 +96,8 @@ setup(name='Hyperion',
                 'hyperion.densities.tests',
                 'hyperion.dust',
                 'hyperion.dust.tests',
-                'hyperion.filters',
-                'hyperion.filters.tests',
+                'hyperion.filter',
+                'hyperion.filter.tests',
                 'hyperion.grid',
                 'hyperion.grid.tests',
                 'hyperion.importers',
